Refactor BK7231Serial, detect protocol type by CRC value
