+++ bk7231tools/serial/base/__init__.py
@@ -1,0 +1,18 @@
+#  Copyright (c) Kuba Szczodrzyński 2024-3-5.
+
+from .data import BK7231SerialData
+from .enums import Bootloader, BootloaderType, ChipType, ProtocolType
+from .interface import BK7231SerialInterface
+from .packets import EraseSize, Packet
+
+__all__ = [
+    "BK7231SerialData",
+    "BK7231SerialInterface",
+    "ChipType",
+    "ProtocolType",
+    "Bootloader",
+    "BootloaderType",
+    "Packet",
+    "EraseSize",
+    "packets",
+]
